python-flask/flask-api/requirements.txt
alembic==1.7.7
astroid==2.11.2
attrs==21.4.0
click==8.1.2
dill==0.3.4
Flask==2.1.1
Flask-CLI==0.4.0
Flask-Cors==3.0.10
Flask-JWT==0.3.2
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.3
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.1
jsonschema==4.4.0
lazy-object-proxy==1.7.1
Mako==1.2.0
MarkupSafe==2.1.1
mccabe==0.7.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
PyJWT==1.4.2
pylint==2.13.5
pyrsistent==0.18.1
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.35
tomli==2.0.1
typing_extensions==4.2.0
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
